End of ml logic

+++ ml_service/src/train_model.py
@@ -1,35 +1,35 @@
-import os
 import pandas as pd
 import joblib
-from sklearn.feature_extraction.text import TfidfVectorizer
+import os
+from src.bert_utils import CodeEmbedder 
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-
-DATA_PATH = os.path.join(BASE_DIR, "..", "data", "code_bug_fix_pairs.csv")
-VECTORIZER_PATH = os.path.join(BASE_DIR, "vectorizer.pkl")
+DATA_PATH = os.path.join(BASE_DIR, "..", "data", "code_bug_fix_pairs_merged.csv")
+DB_VECTORS_PATH = os.path.join(BASE_DIR, "db_vectors.pkl")
 DB_PATH = os.path.join(BASE_DIR, "fix_database.pkl")
 
 def train():
+    print(f"Loading data from: {DATA_PATH}")
+    if not os.path.exists(DATA_PATH):
+        print("CSV file not found.")
+        return
+
     df = pd.read_csv(DATA_PATH)
-    df = df[["buggy_code", "fixed_code", "commit_message"]].dropna()
+    df = df.dropna(subset=["buggy_code"])
+    
+    df["train_text"] = df["buggy_code"].str.strip()
 
     if df.empty:
-            print("Dataset is empty, skipping training.")
-            return
+        print("Dataset is empty.")
+        return
+
+    embedder = CodeEmbedder()
     
-    vectorizer = TfidfVectorizer(
-    analyzer="char_wb",  # "char_wb" учитывает границы слов (лучше для кода)
-    ngram_range=(2, 4),  # Берем кусочки по 2-4 символа (больше шансов совпасть)
-    min_df=1,            # Учитывать даже редкие ошибки
-    max_features=5000
-)
+    vectors = embedder.get_embeddings(df["train_text"].tolist())
 
-    X = vectorizer.fit_transform(df["buggy_code"])
-
-    joblib.dump(vectorizer, VECTORIZER_PATH)
     joblib.dump(df, DB_PATH)
+    joblib.dump(vectors, DB_VECTORS_PATH)
 
-    print("ML model trained: vectorizer + fix database saved")
 
 if __name__ == "__main__":
     train()
